Moving linked image handling into Carousel.images
In tests using `.find_next()` instead of `.find_all("img")` since the
img tag is expected to be the first child element.

In tests using `.find_next()` instead of `.item` since the latter
doesn't guarantee the returned item to be an element, can be a newline
which we want to ignore.