New feature: buttons_on_top
This solves a UX problem where a carousel containing images with
different aspect ratios or different caption lengths cause the prev/next
controls to shift when sliding to the next/prev image. This happens
because the controls are set to be in the middle of the entire carousel,
so taller images and longer captions extend the height of the carousel
and affect the button placements.

This option pins the controls to the top of the carousel so they no
longer move when the above situation occurs. The indicators are also
pinned to the top since they too move around as they're pinned to the
bottom of the carousel by default.