Implementing hacky Sphinx multi-theme support.
Fork the Python process when Sphinx is parsing conf.py and change
`html_theme` as well as `outdir` in the child process. The first theme
in the list is the main theme that will be used for the usual
`_build/html/index.html` page, and subsequent themes will be built into
`_build/html/theme_alabaster/index.html` as an example.

For now building with alabaster and classic as the subsequent themes
without inspecting them for CSS issues with sphinx-carousel, that will
be taken care of in the future.

Maybe one day MultiTheme will be a separate library but for now I'm
tossing it into this project until I need to use it for another project.
Since this is kind of hacky and the forking part doesn't work with
`sphinx.testing` I'm disabling some coverage from most of the
`unrelated.py` file. May need to reimplement in the future and make it
less hacky, or suggest additional APIs in sphinx-doc or something.